Fixed problem with removing last existing row & added extra test check #7548

+++ wrappers/vue/src/HotTable.vue
@@ -156,7 +156,7 @@
             this.hotInstance.rowIndexMapper.insertIndexes(oldDataLength - 1, data.length - oldDataLength);
           }
 
-          if (isColumnModificationAllowed) {
+          if (isColumnModificationAllowed && data.length !== 0) {
             if (columnsToRemove.length > 0) {
               this.hotInstance.columnIndexMapper.removeIndexes(columnsToRemove);
